scripts/requirements.txt
anthropic>=0.18.0
mammoth>=1.6.0
python-docx>=1.0.0
